# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  request.DailyRewardMuti:
    properties:
      type:
        type: integer
    type: object
...
  request.PlayerLevelRewardCollectRequest:
    properties:
      level:
        type: integer
    type: object
...
  request.UpdatePlayer:
    properties:
      avatar:
        type: integer
      username:
        type: string
    type: object
  request.UpdatePlayerSettingsRequest:
    properties:
      friendRequests:
        type: boolean
      joystick_size:
        type: number
      language:
        type: string
      music:
        type: number
      notifications:
        type: boolean
      sound:
        type: number
      vibration:
        type: boolean
      voicePack:
        type: boolean
    type: object
  response.Status:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  response.Success:
    properties:
      status:
        $ref: '#/definitions/response.Status'
    type: object
info:
  contact: {}
  description: This is the api documentation of survival party game
  title: Survival Party
  version: "1.0"
paths:
...
  /collect-daily-rewards:
    put:
      consumes:
      - application/json
      description: Collect the daily reward of the  player
      parameters:
      - description: Player Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Reward Details
        in: body
        name: rewardType
        required: true
        schema:
          $ref: '#/definitions/request.DailyRewardMuti'
      produces:
      - application/json
      responses:
        "200":
          description: Sucess
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Collect the daily reward of the  player
      tags:
      - DailyReward
  /daily-rewards:
    get:
      consumes:
      - application/json
      description: Gets the list of daily reward for the  player
      parameters:
      - description: Player Access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sucess
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Gets the list of daily reward for the  player
      tags:
      - DailyReward
...
  /get-level-rewards:
    get:
      consumes:
      - application/json
      description: Gets the rewards according to player level
      parameters:
      - description: Player Access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Gets reward list
      tags:
      - Player
...
  /get-settings:
    get:
      consumes:
      - application/json
      description: Gets the current settings of that player
      parameters:
      - description: Player Access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sucess
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Gets the settings
      tags:
      - Player
  /level-reward-collect:
    post:
      consumes:
      - application/json
      description: Collects the reward for a level of that player
      parameters:
      - description: Player Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Player Details
        in: body
        name: loginDetails
        required: true
        schema:
          $ref: '#/definitions/request.PlayerLevelRewardCollectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Collects Reward
      tags:
      - Player
  /name-time-left:
    get:
      consumes:
      - application/json
      description: Get time left for name change
      parameters:
      - description: Player Access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Get time left for name change
      tags:
      - Player
...
  /stats:
    get:
      consumes:
      - application/json
      description: Get the player game stats
      parameters:
      - description: Player Access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Get player game stats
      tags:
      - Player
...
  /update-settings:
    put:
      consumes:
      - application/json
      description: Updates the game settings of that player
      parameters:
      - description: Player Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Player Settings
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/request.UpdatePlayerSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Sucess
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Updates setting
      tags:
      - Player
  /user-data:
    get:
      consumes:
      - application/json
      description: Get game stats of ther players
      parameters:
      - description: Id of the other player
        in: query
        name: playerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Get game stats of ther players
      tags:
      - Player
    put:
      consumes:
      - application/json
      description: Updates player info like username and avatar
      parameters:
      - description: Player Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Player Details
        in: body
        name: loginDetails
        required: true
        schema:
          $ref: '#/definitions/request.UpdatePlayer'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Updates player info
      tags:
      - Player
...
swagger: "2.0"